Update LINE file.cs

+++ folder1/folder2/file.cs
@@ -1,2 +1,3 @@
 Console.WriteLine("How are you? am fine");
-int a = 0;
+double a = 0;
+//UPDATE
